requirements.txt
pydantic-ai
psycopg2-binary
pgvector
openai
python-dotenv
pytest
